Move queries to different file.
